GG-32437 Create a JMX metric to show realtime sql engine memory utilization change freeMem and maxMem to be sourced directly from QueryMemoryManager
