+++ modules/indexing/src/test/java/org/apache/ignite/internal/metric/SqlStatisticsAbstractTest.java
@@ -225,4 +225,32 @@ public class SqlStatisticsAbstractTest extends GridCommonAbstractTest {
             throw new RuntimeException("Fail the query.");
         }
     }
+
+    /**
+     * Functional interface to validate memory metrics values.
+     */
+    protected static interface MemValidator {
+        /**
+         *
+         * @param free freeMem metric value.
+         * @param max maxMem metric value.
+         */
+        void validate(long free, long max);
+    }
+
+    /**
+     * This callback validates that some memory is reserved.
+     */
+    protected static final MemValidator MEMORY_IS_USED = (freeMem, maxMem) -> {
+        if (freeMem == maxMem)
+            fail("Expected some memory reserved.");
+    };
+
+    /**
+     * This callback validates that no "sql" memory is reserved.
+     */
+    protected static final MemValidator MEMORY_IS_FREE = (freeMem, maxMem) -> {
+        if (freeMem < maxMem)
+            fail(String.format("Expected no memory reserved: [freeMem=%d, maxMem=%d]", freeMem, maxMem));
+    };
 }
